+++ assets/Colors.tsx
@@ -1,14 +1,11 @@
+// eslint-disable-next-line import/prefer-default-export
 export const Colors = {
-    navbarPink: '#F3C8CC',
-    brandBlue: '#0F2536',
-    lightBlue: '#9FC5E5',
-    brandRed: '#C92C3C',
-    textBlack: '#2B2B2B',
-    brandGray: '#6A6A6A',
-    lightGray: '#E7E7E7',  
-    white: '#FFFFFF',
-}
-        
-        
-
-    
+  navbarPink: '#F3C8CC',
+  brandBlue: '#0F2536',
+  lightBlue: '#9FC5E5',
+  brandRed: '#C92C3C',
+  textBlack: '#2B2B2B',
+  brandGray: '#6A6A6A',
+  lightGray: '#E7E7E7',
+  white: '#FFFFFF',
+};
