+++ screens/Documents/Documents.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { TextRegular } from '../../assets/fonts/Fonts';
+import { PageContainer } from '../styles';
+
+const UploadScreen = () => {
+  return (
+    <PageContainer>
+      <TextRegular>Upload your necessary documents!</TextRegular>
+    </PageContainer>
+  );
+};
+
+export default UploadScreen;
