+++ screens/Forms/FormsScreen.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { TextRegular } from '../../assets/fonts/Fonts';
+import { PageContainer } from '../styles';
+
+const FormsScreen = () => {
+  return (
+    <PageContainer>
+      <TextRegular>Fill out your intake form!</TextRegular>
+    </PageContainer>
+  );
+};
+
+export default FormsScreen;
