+++ screens/Home/HomeScreen.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { TextRegular } from '../../assets/fonts/Fonts';
+import { PageContainer } from '../styles';
+
+const HomeScreen = () => {
+  return (
+    <PageContainer>
+      <TextRegular>Welcome to the homepage!</TextRegular>
+    </PageContainer>
+  );
+};
+
+export default HomeScreen;
