+++ screens/styles.ts
@@ -1,0 +1,8 @@
+/* eslint-disable import/prefer-default-export */
+import styled from 'styled-components/native';
+
+export const PageContainer = styled.View`
+  margin: 5%;
+  justify-content: 'center';
+  align-items: 'center';
+`;
